% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clogitR.R, R/clogitRFunctions.R
\docType{package}
\name{clogitR}
\alias{clogitR-package}
\alias{clogitR}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{	Conditional Logistic Regression with Optional Concordant pairs 
}\if{html}{\out{</div>}}}
\usage{
clogitR(formula, data, use_concordant_pairs = TRUE, try_to_stableize = TRUE)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data.frame or data.table containing the variables.}

\item{use_concordant_pairs}{Flag indicating whether to include the concordant pairs and reservoir in the model.}

\item{try_to_stableize}{Flag indicating whether the model should decompose the discordant pairs to increase stability if the model is unstable.}
}
\value{
A new \code{clogitRClass} object.
}
\description{
clogitR
}
\details{
Fits a conditional logistic regression model to incidence data. Allows for use of the concordant pairs in the fitting.
}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{paper to go here
}\if{html}{\out{</div>}}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	Jacob Tennenbaum \email{Jacob.Tennenbaum51@qmail.cuny.edu}
}\if{html}{\out{</div>}}
}
\keyword{Conditional}
\keyword{Logistic}
\keyword{Regression}
