% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalLogisticRegressionClass.R
\name{clogitRClass}
\alias{clogitRClass}
\title{Conditional Logistic Regression}
\description{
\code{clogitRClass} is an R6 class for fitting conditional logistic regression
models using stratified case-control data. This is a skeleton implementation
containing structure and documentation, but no estimation routine.
}
\details{
The class accepts a formula interface (\code{y ~ x1 + x2 + strata(stratum)})
and a \code{data.frame} or \code{data.table}. Internally, it parses the formula,
extracts variables, and prepares the data for model fitting.
}
\examples{
\dontrun{
data <- data.frame(
  y = c(1, 0, 1, 0),
  x = c(2.3, 1.5, 3.1, 2.9),
  stratum = c(1, 1, 2, 2)
)

fit <- clogitR$new(y ~ x + strata(stratum), data = data)
fit$summary()
}

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{formula}}{The model formula supplied by the user.}

\item{\code{data_name}}{The name of the dataset provided to the model.}

\item{\code{data}}{The dataset provided to the model.}

\item{\code{model_frame}}{The model frame generated from the formula and data.}

\item{\code{terms}}{Model terms object from the formula.}

\item{\code{use_concordant_pairs}}{Indicator to fit the concordant pairs.}

\item{\code{try_to_stableize}}{Indicator to decompose the discordant pairs the pairs if the model is unstable.}

\item{\code{X_reservoir_concordant}}{X rows marked with 0 in the strata or where part of the concordat pairs.}

\item{\code{y_reservoir_concordant}}{y values marked with 0 in the strata or where part of the concordat pairs.}

\item{\code{X_diffs_discordant}}{Difference of X rows in the strata. If treatment was indicated in the formula the difference is T - C.}

\item{\code{y_diffs_discordant}}{Difference of y values in the strata. If treatment was indicated in the formula the difference is T - C.}

\item{\code{coefficients_and_variances}}{Table containing the coefficients and variances of each covariate.}

\item{\code{droped_reservoir_concordant}}{Flag indicating whether the reservoir and concordant data was doped.}

\item{\code{droped_discordant}}{Flag indicating whether the discordant data was doped.}

\item{\code{converged_reservoir_concordant}}{Flag indicating whether the reservoir and concordant model converged.}

\item{\code{converged_discordant}}{Flag indicating whether the discordant model converged.
Initialize a new clogitRClass object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clogitRClass-new}{\code{clogitRClass$new()}}
\item \href{#method-clogitRClass-fit}{\code{clogitRClass$fit()}}
\item \href{#method-clogitRClass-summary}{\code{clogitRClass$summary()}}
\item \href{#method-clogitRClass-clone}{\code{clogitRClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clogitRClass-new"></a>}}
\if{latex}{\out{\hypertarget{method-clogitRClass-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clogitRClass$new(
  formula,
  data,
  use_concordant_pairs = TRUE,
  try_to_stableize = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formula}}{A formula specifying the model.}

\item{\code{data}}{A data.frame or data.table containing the variables.}

\item{\code{use_concordant_pairs}}{Flag indicating whether to include the concordant pairs and reservoir in the model.}

\item{\code{try_to_stableize}}{Flag indicating whether the model should decompose the discordant pairs to increase stability if the model is unstable.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{clogitRClass} object.
Fit the model
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clogitRClass-fit"></a>}}
\if{latex}{\out{\hypertarget{method-clogitRClass-fit}{}}}
\subsection{Method \code{fit()}}{
Placeholder function for estimation. Currently not implemented.
Summarize the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clogitRClass$fit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clogitRClass-summary"></a>}}
\if{latex}{\out{\hypertarget{method-clogitRClass-summary}{}}}
\subsection{Method \code{summary()}}{
Placeholder summary function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clogitRClass$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clogitRClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clogitRClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{clogitRClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
