% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastPairwiseClogit.R
\name{fastClogit}
\alias{fastClogit}
\title{Fast Conditional logistic regression with prepared data without any checks}
\usage{
fastClogit(
  discordant_Xdiffs = NULL,
  discordant_ydiffs = NULL,
  discordnat_Treatmentdiffs = NULL,
  concordnat_X = NULL,
  concordnat_y = NULL,
  concordnat_Treatment = NULL
)
}
\arguments{
\item{discordant_Xdiffs}{A fully prepared data matrix of the differences of the discordant matched pairs.}

\item{discordant_ydiffs}{A fully prepared binary (-1,1) vector of the differences in response the discordant matched pairs.}

\item{discordnat_Treatmentdiffs}{A fully prepared binary (-1,1) vector of the differences in treatment the discordant matched pairs.}

\item{concordnat_X}{A fully prepared data matrix of the concordant matched pairs or reservoir.}

\item{concordnat_y}{A fully prepared binary (0,1) vector of the response the concordant matched pairs or reservoir.}

\item{concordnat_Treatment}{A fully prepared binary (0,1) vector of the treatment the concordant matched pairs or reservoir.}
}
\value{
A list containing the estimated treatment effect and squared standard error for each of the applicable models (discordant, concordant, mixed).
}
\description{
Computes the treatment effect and its squared standard error for the discordant and concordant pairs.
It then returns a list containing theses values as well as the mixed estimates.

This is the fast, very bare-bones, version of the model, primary meant for power simulations or other situations where
the overhead of object creation is too cumbersome. As such there are no checks, validations, or warnings, in this model.
It will break if there are not enough renovates, or the length of vectors don't match, and doesn't make any attempt to
stabilize potentially unstable results. USE AT YOUR OWN RISK.

For the standard model, see \code{clogitR}.
}
