% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clogitR.R, R/clogitRFunctions.R
\docType{package}
\name{clogitR}
\alias{clogitR-package}
\alias{clogitR}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{	Conditional Logistic Regression with Optional Concordant pairs 
}\if{html}{\out{</div>}}}
\usage{
clogitR(
  response = NULL,
  data = NULL,
  treatment = NULL,
  strata = NULL,
  use_concordant_pairs = TRUE,
  try_to_stableize = TRUE
)
}
\arguments{
\item{response}{A binary (0,1) vector containing the response of each subject. HI}

\item{data}{A data.frame, data.table, or model.matrix containing the variables.}

\item{treatment}{A binary (0,1) vector containing weather each subject is treatment or control.}

\item{strata}{Numeric vector indexing the matched pairs and the reservoir. Entrees indexed with zero, are added to the reservoir. Entrees indexed with the same number are part of the same strata.}

\item{use_concordant_pairs}{Flag indicating whether to include the concordant pairs and reservoir in the model.}

\item{try_to_stableize}{Flag indicating whether the model should decompose the discordant pairs to increase stability if the model is unstable.}
}
\value{
A new \code{clogitRClass} object.
}
\description{
clogitR
}
\details{
Fits a conditional logistic regression model to incidence data. Allows for use of the concordant pairs in the fitting.
}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{paper to go here
}\if{html}{\out{</div>}}
}
\author{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	Jacob Tennenbaum \email{Jacob.Tennenbaum51@qmail.cuny.edu}
}\if{html}{\out{</div>}}
}
\keyword{Conditional}
\keyword{Logistic}
\keyword{Regression}
